feat(playground): updating playground styles
