+++ packages/playground/src/pages/main-layout.tsx
@@ -45,10 +45,13 @@ export const MainLayout = () => {
 				}
 			>
 				<GlobalNav />
-				<SidebarInset>
+				<SidebarInset
+					data-testId="main-layout"
+					className="m-0! shadow-none"
+				>
 					<div
 						data-testId="main-layout"
-						className="h-[calc(100vh-theme(space.2))] w-full overflow-hidden"
+						className="h-screen w-full overflow-hidden"
 						style={{
 							background:
 								"linear-gradient(180deg, #FCFCFC 58.78%, #F6F7FF 81.97%, #F1F8FF 94.04%), var(--base-secondary-background, #FFF)",
